LPS-102711 Moves details to its own screen
